+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 2.9.0 (TBD)
 
+- Feature: The traffic-manager can now configure all clients that connect through the `client:` map in
+  the `values.yaml` file.
 - Feature: The traffic-manager version is now included in the output from the `telepresence version` command.
 
 ### 2.8.5 (November 2, 2022)